Enhance the handling of search fields
- Allow customizing the filter dropdown menu (field name and label)
- Pre-select currently active filter and show filter value

To customize the available search fields, supply a list property on the
CrudRestController which contains tuples of (field, name) or strings
with field = name which will be used for displaying the available
filters. If you specify True, the previous mechanism of automagically
generating the fields from the table headers will be used (default).
If you specify any False value, there will be no search fields displayed
at all.